use defaultEventformat when format undefined, aligned with articles
