+++ common/data/defaultValues.ts
@@ -1,0 +1,18 @@
+// used for display purpose only, when the prismic data is undefined
+export const defaultArticleFormat = {
+  type: "ArticleFormat",
+  id: "W7TfJRAAAJ1D0eLK",
+  label: "Article",
+};
+
+export const defaultEventFormat = {
+  type: "EventFormat",
+  id: "dfc2b7f9-c362-47da-9644-b0f98212ccaa",
+  label: "Event",
+};
+
+export const onlineLocation = {
+  type: "EventLocation",
+  id: "ef04c8e3-26be-4fbc-9bef-f52589ebc56c",
+  label: "Online",
+};
